[float-x-y] Отправляю float-координаты

+++ src/Entity/Snake.php
@@ -6,8 +6,8 @@ class Snake
 {
     public function __construct(
         private int       $id,
-        private int       $headX,
-        private int       $headY,
+        private float       $headX,
+        private float       $headY,
         private array     $bodyParts,
         private int       $radius,
         private Direction $direction,
@@ -23,12 +23,12 @@ class Snake
         return $this->id;
     }
 
-    public function getHeadX(): int
+    public function getHeadX(): float
     {
         return $this->headX;
     }
 
-    public function getHeadY(): int
+    public function getHeadY(): float
     {
         return $this->headY;
     }
@@ -71,12 +71,12 @@ class Snake
         return count($this->bodyParts) + 1; // + head
     }
 
-    public function setHeadX(int $headX): void
+    public function setHeadX(float $headX): void
     {
         $this->headX = $headX;
     }
 
-    public function setHeadY(int $headY): void
+    public function setHeadY(float $headY): void
     {
         $this->headY = $headY;
     }
@@ -111,7 +111,7 @@ class Snake
         $this->score += $add;
     }
 
-    public function addBodyPart(int $x, int $y, string $color): void
+    public function addBodyPart(float $x, float $y, string $color): void
     {
         $this->bodyParts[] = new BodyPart($x, $y,
                                           $this->getRadius(),
